(PCP-657) Drop ctrun on Solaris
Before this patch we used the `ctrun` command on Solaris to execute external
module actions. This was necessary for the external action processes to be
able to survive possible restart of the pxp-agent service on that platform.
Unfortunately the use of the `ctrun` command introduced a race into the
pxp-agent's non-blocking action processes book keeping logic. The problem
was that pxp-agent believed the `ctrun` command was the process executing
the action, so when it exited during the pxp-agent restart, the new pxp-agent
instance assumed (upon detection the process was no longer running) that
the action had finished, but if that wasn't the case (if the real action
process was still executing) the action's status was incorrectly reported as
`unknown`.
To solve the problem the `ctrun`'s functionality of executing child processes
such that they can survive their parents was added to `leatherman::execution`
and can be used by specifying `create_detached_process` execution option,
which is what we take advantage of in this patch.