(PA-777) Implement puppetserver reload
This commit update how the puppetserver service is restarted in the
pxp-agent testing pipelines. Recently we added functionality into
puppetserver to allow for faster reload times. By taking advantage of
this functionality, we should be able to reduce the amount of time our
pxp-agent test pipelines take to execute.

This commit also sets the `restart_when_done beaker` option to false globally
in AIO, which requires Beaker 2.27.0 to work correctly and is already
the minimum beaker version we depend on for acceptance.