(PCP-736) Don't change umask
Previosly pxp-agent set its umask to 020 but only when it was daemonizing
itself. I.e. it did not do that when executed from systemd/launchd.
This made it difficult for users to configure the umask for the puppet
runs executed by pxp-agent.
So in this patch we drop the setting of the umask in the pxp-agent to make
it possible for users to override the umask in an init script and/or systemd
launchd service file.