(PCP-742) Return metrics with pxp-module-puppet result
Previously, we returned on the transaction_uuid of the report and relied
on orchestrator to use that to retrieve metrics about what happened.
Because the report may not necessarily have been stored in PuppetDB by
the time that data was looked up, orchestrator could fail to find
metrics immediately after the run finished. To get around that, we now
return the metrics as part of the result, rather than requiring it to
separately query the data we already have.