(maint) Rewrite kill test for PUP-6675
The test around Puppet starting after a previous Puppet run has
completed previously relied on Windows behavior where the Puppet process
would exit completely even though a child process is still running.

PUP-6675 changes Puppet to use pipes for stdout from invoking a child
process; if Puppet is terminated while the child process is still
running, it will still have a valid process handle until the child
process exits (and closes the shared pipe). This is explained in
https://msdn.microsoft.com/en-us/library/windows/desktop/ms686714(v=vs.85).aspx.

Puppet terminating while a direct child process is still running is not
normal behavior, so modify the test to not rely on it.

[skip ci]