(PCP-699) Accept and pass job-id in pxp-module-puppet
In order to track the execution of a job all the way through the agent
to PuppetDB, we now have a --job-id flag that can be passed to `puppet
agent`. That flag will cause the provided job id to be added to the
report that gets stored in PuppetDB, making it easy to determine whether
a report was associated with a job.

We now optionally accept a "job-id" input to pxp-module-puppet and will
pass that along to the agent. This is provided as a top-level input
rather than a flag for backward compatibility (so old agents will ignore
the option rather than fail), and because it properly represents
metadata about the run rather than being an ordinary flag.