(PCP-754) Support stopping service when the binary path changes
From pxp-agent 1.5.3 to 1.6.0, pxp-agent used a versioned binary with
pxp-agent symlinked to it. When upgrading, the binary used to start the
pxp-agent service would change. On SUSE, this prevented restarting the
service after upgrade because killproc would not kill the process
started with the previous binary. Use kill instead - with appropriate
guards to check it was the expected service process - to ensure we can
restart the service after upgrading from one of those versions.

[skip ci]