(PCP-759) Enable EC2 master install
This commit updates the acceptance install pre-suite to allow installing
an ec2 master dynamically on a el based system without intervention.
Due to the expectations of the nightly puppetserver repo, the nightly
puppet-agent is installed first and then over-written with the
puppet-agent version under test. The specification of a SERVER_VERSION
is also supported.

A `LEIN_PROFILE` constant has been added to switch between profiles
based on whether the SUT has access to the `puppetlabs.net` network
or not to enable the installation of pcp-broker on the master.