(PCP-758) Refactor find_task, ensure found matches task name
Pulls find_task into a class, with a map for looking up extensions to
make it easier to add new ones.

Previously the directory iterator would check all files, so the task
file that's found wouldn't necessarily match the task name. Fix it so we
only consider files matching the task name.