(PCP-633) Report running if action thread is still active
During a non-blocking request, if a status query is recieved after
sending the provisional response but before the action thread starts a
subprocess and writes the PID file it will report status "unknown". In
PCP-767, we'll be adding a new module that downloads a file before
starting a subprocess; if that download takes awhile, we would be in an
"unknown" state for an inappropriate amount of time.

Report "running" if the action thread is still active rather than
"unknown", as we know the action is still running.