(PCP-769) Download tasks from master-uri
With this commit, there are now two cases to consider when retrieving
a cached task file. If the task file exists and its sha matches the
provided sha sent by orchestrator, then the file is executed (note
that the first file sent by orchestrator is assumed to be the
executable) with owner/group read, owner write, and owner/group
execute permissions for Linux; and parent-directory permissions for
windows. All task-files will now reside in the <tasks-cache>/<sha256>
directory.

If the task file does not exist, then it is downloaded using the first
provided master-uri. The sha of the downloaded file is checked with
the orchestrator-provided sha. If the shas match, then the task
file is executed (with owner/group read, owner write and owner/group
execute permissions). Otherwise, a PXP error object is returned
indicating a version conflict.