(PCP-776) Detect disabled/running Puppet via lock files
`pxp-module-puppet` can't detect that a run failed - because Puppet was
disabled or because Puppet was already running - when Puppet's output is
translated. What we currently try to do is
- if Puppet's disabled, return an error message to that effect
- if Puppet's running, wait until it stops and start a new run

Switch to checking for the lockfiles that Puppet uses to determine
disabled or running, rather than parsing the output. This makes the
implementation language-agnostic.