(maint) Improve test debugibility and failure handling
When an action finishes, it always returns an exitcode. stdout and
stderr are only included if they have contents. Identify completion by
the presence of exitcode so we can finish early, rather than timing out
if a failure means stdout is empty.

Also include debug logging when Beaker is configured to use debug
output.

[skip ci]