(PCP-786) Send error response to status if found
If an execution error is found in the metadata of a non-blocking action,
send a response with `{"_error": {"kind":
"puppetlabs.pxp-agent/execution-error", "msg": <msg>, "details" {}}}` on
stdout. This should only happen if an error occurs while invoking an
action that would result in sending an `rpc_error_message`, and is done
to make sure the same error message can be retrieved by a status request
if the `rpc_error_message` is lost.