(maint) Fix local tmp dir creation
Test setup depends on a local tmp dir existing. We create it when
generating a hosts file, and used to create it when installing the
puppet-agent package. We no longer create it for install, but still use
it for testing and the hosts file. Ensure it's always created in the
Rakefile so it's available during setup.

[skip ci]