(maint) Ensure umask can run on master
Prior to this commit, we were unable to run this test on the master.
This was because we were using the environment name to represent a file
that we created on each agent. Unfortunately, this file clashed with the
actual environment on the master, so this test failed because the file
it was looking for was actually a directory. This commit removes that
collision, allowing us to run this test even if the master has also been
classified as an agent.