(maint) Remove check for pcp_broker_instance
The code is all written to assume that pcp_broker_instance is 0 if it
isn't set. With the switch to using beaker subcommands to run acceptance
tests, we aren't preserving data between runs (i.e., between the
pre-suite and test runs). This means that there is some data that is
lost, for example `master[:pcp_broker_instance]`. However, we shouldn't
need this data anyway. This commit removes the assertions around how
pcp_broker_instance is set.