(PCP-849) Only pass supported parameters to powershell tasks
Powershell tasks will fail if they receive parameters that they don't
explicitly declare. Since we want to pass some metaparams regardless of
whether the task needs them, we need to ensure that we exclude those
parameters for powershell tasks that don't understand them.

This follows Bolt's method for restricting the parameters passed to the
task, by inspecting the command and selecting only the ones it expects.