(PCP-851) Download files required by task
In order to support multi-file tasks pxp-agent should find all files required for the chosen implementation for the task and download them from puppet-server. When a NonBlocking request is made the _installdir is built in the pxp-agent spool dir. When a Blocking request is recieved the installdir is built in the task-cache. Blocking requests to run a task should only be expected in tests. The _installdir location is passed to the task as metadata.