(PA-2219) Add support for Fedora 30
Due to Fedora removing its 'service' command in release 30, we need to
better handle checking the service provider in the acceptance tests.

Previous implementation assumed the 'service' command is always
present, however Fedora 30 and future releases will not include
this command/package, as Sys-V services are deprecated in favor
of systemd.