(PA-2267) Update pxp tests to check /proc/1/exe for systemd
Previous to this commit, two tests: run_puppet_umask and
multiple_pxp_agent_daemon were attempting to use the 'service' command to check
for systemd. This will not work for newer fedora versions since sysvinit was
removed and the 'service' command went with it.

Thus we check if the file that /proc/1/exe points to is systemd