(PCP-868) Add download_file module
This commit adds a new boltmodule child: DownloadFile. This pxp module simply
downloads a file and returns, with no calls to external processes.

Updates were required to make downloadFileFromMaster actually work correctly
outside of task.cc. Specifically: two other functions needed to be moved from
task.cc to bolt_helpers.cc: createDir and createCacheDir

Since DownloadFile does not need external processes: the class implements its'
own version of callAction, which does all the processing. To make the class
continue to work as a BoltModule child an empty definition of
buildCommandObject was added to the class (in the header file).