(PCP-868) Add download_file module
This commit adds a new boltmodule child: DownloadFile. This pxp module simply
downloads a file and returns, with no calls to external processes.

Since DownloadFile does not need external processes: the class implements its'
own version of callAction, which does all the processing. To make the class
continue to work as a BoltModule child an empty definition of
buildCommandObject was added to the class (in the header file).