(PCP-874) Update download_file to work with multiple files
This commit updates the download_file module to work with multiple files. The
schema already defined that the PCP message should be an array of files, but
previously the module wouldn't operate on anything other than the first file.

This commit updates the module to actually iterate over all the files.