(PCP-874) Update download_file to work with symlinks
This commit updates the download_file pxp module with implementation for the
"symlink" file type. A new entry has been added to the schema for the source
where the link will point to: "link_source". This field is expected to be
empty for directory and file types.

"download_file" actions will fail if the symlink to be created already exists
and is _not_ a symlink (i.e. the destination already exists and is either a
file or directory).