(PCP-869) Move findExecutable function to bolt_helpers
Previously to this commit, the task module contained a helper function:
findExecutableAndArguments() which would identify the executable to use for
windows platforms (where #! is unavailbale).

the new run_script module will need to use the same helper, so this refactor
commit just moves that function (and it's associated constant
BUILTIN_INTERPRETERS) to bolt_helpers for use in both tasks and the new
run_script module.