(PCP-869) Refactor argument tokenizing to utility function
Previously, the command pxp module implemented functionality to tokenize the
command line string in to tokens that could be split between the command to
run and the arguments to pass to leatherman::execute.

The functionality for tokenizing all the words in the command string will be
useful in the new script module when it needs to tokenize all the arguments.

This commit refactors the original functionality out of the command module and
in to a function in bolt_helpers.