(maint) refactor buildCommandObject
Previous to this commit bolt modules each implemented a virtual function:
buildCommandObject to create the command that was to be run with
leatherman::execute. When PXP actually made a request the parent BoltModule
class had an implementation of callAction (a PXP Module class function) that
would call buildCommandObject to create the command, and then the BoltModule
implementation would be what actually executed the functionality.

However, this meant that child classes only controlled creating the command
object, rather than actual execution. The problem with this approach is if
anything outside of building the command object failed (i.e. if a file
download needed to be executed before the command but it failed) the
implementation of buildCommandObject had no way to bubble up that failure and
correctly produce a failure state that is readable.

This commit refactors the base BoltModule class so it no longer contains the
buildCommandObject virtual function, and no longer implements callAction.
Instead, the child classes should implement callAction, and a new function:
invokeCommand is available to the child classes as the last thing to call in
callAction.

The idea behind the new implementation is the BoltModule child classes should
still be building command objects, but it should be inside the callAction
function where they have total control of any exception handling and failure
reporting.