(maint) Refactor failure_response to parent BoltModule
This commit moves the failure_response function from the download_file module
to the parent class BoltModule. The functionality for failure_response is
re-usable in any bolt module where there is a failure outside of running a
command (i.e. if there's a processing or download error before executing the
command).