(maint) Refactor tests for splitArguments
Previously, testing for splitArguments functionality was done in the command
module tests. Now that the buildCommandObject no longer exists testing for
splitArguments must be done somewhere else. This commit moves spec testing for
that functionality out of command_test and in to a new test file
bolt_helpers_test. The new tests should cover the same ground that the old
command tests did, but specifically only calling splitArguments.

Additionally, new tests were added to script_test that execute testing of
arguments and files with spaces.