(maint) rename download_file to file
The download_file module in pxp-agent does not match the same naming scheme
that all the other modules used. All other modules are named for the noun they
act on - script, command, task - and then the modules contain an action for
the verb. By this scheme the download module should be called: "file" and it
should have one action "download".