(PCP-875)(PCP-891) Add apply module to pxp-agent
This commit adds a new internal apply module. The apply module has a single `apply` action which will apply a catalog on an agent. The result of an apply is a JSON report of the puppet run. Currenly the ruby shim is managed as a HEREDOC in the apply.cc source. This should be managed as a shipped file in pxp-agent packaging. We will do that work in subsequent tickets.

This commit updates the ruby shim and the apply module C++ code to enable downloading plugins in to the temporary environment the catalog is applied in. Currently, the C++ code is responsible for ensuring a cache directory is available for the shim to download plugins in to. The shim is responsible for ensuring the plugins are up to date and copying them in to the expected location in the temporary environment.

We pass through the connection information from the pxp-agent through to the shim. We use this for pluginsync and then clear all setting before applying the compiled catalog to ensure isolation. Currently PXP does not honor CRL. We need to enable that in pxp-agent and then we can pass that through to the shim as well.

In order to download plugins from puppetserver with puppet in the apply shim we need CRL. This commit makes CRL a required parameter for pxp-agent for use with the apply module. Plumbing in the agent to honor the CRL in the other built in modules as well as connecting to the PCP broker will be subsequently implemented. It is tracked with https://tickets.puppetlabs.com/browse/PCP-893