(maint) Add bolt plans for building/rebuild pxp-agent
This commit adds a new Boltdir and two plans for building and rebuilding
pxp-agent. Both plans take a host as input where an agent is installed,
and will update the host's pxp-agent binary (and pxp-agent modules dir)
with changes built using the files currently in the local checkout of
pxp-agent.

The plans to use are:

* pxp_dev::build
* pxp_dev::rebuild