(PCP-894) Add `prep` action to `apply` module
This commit adds a `prep` action to the `apply` module. This includes the addition of input/output schemas for the new request (notably the only required input is the `environment`). This action in facilitated by the same ruby shim. Logic in both the `Apply::buildCommandObject` and ruby shim code handles the (slight) differences between the `apply` and `prep` actions. Specifically the setup for pluginsync and gathering facts is the same between the actions. The only difference is that the `prep` action just returns facts while the `apply` action continues on and applies the catalog.