(PCP-896) Add bolt environment server setup
This commit adds a fixture directory for a 'bolt' environment.  This
directory contains the bolt tasks, scripts, and download files used for
various bolt tests.

An additional setup step is added to rsync this environment to the
puppetserver.  This will provide the pxp-agent tests the necessary
puppetserver configuration without having to setup and restart the
puppetserver in an ad-hoc basic per test.

The server setup step is also added to the 'aio' options file to ensure
that it is executed as part of the setup suite for an 'aio'
installation.