(PE-29137) Make use of Curl threadsafe
The use of the Leatherman Curl client is not thread safe (https://github.com/puppetlabs/leatherman/blame/master/curl/inc/leatherman/curl/client.hpp#L271) when module actions that use the client to dowload content from puppetserver were run concurrently a curl error was raised and the action would fail. This comit synchronizes the use of the curl client accross and within modules. Specifically the helper methods that implement curl were moved to the ModuleCacheDir class where a mutex is maintained to synchonize the use of the curl client. The ModuleCacheDir class was chosen because it is instantiated in the RequestProcesser and a pointer to the object is already passed to modules that need to download content from the master. This ensures that the mutex synchonizes the use of the client both between and within modules.