(PE-28827) Handle max-message-size agent side
When a message exceeds the max-message-size parameter in pcp-broker the broker continuously asks the agent for a response, which continues to fail and cause hanging behaviour when executing an action that produces results with large messages. This commit adds the max-message-size setting to the pxp-agent which is intended to be managed in sync with the same setting in pcp-broker so that messages that exceed the max message size are not attempted to be sent to the broker.