(PE-28827) Update max-message-size test with fixes
This commit updates the max-message-size test and the test utils in the
following ways:

1. The test itself is updated to change execution such that a large file is
created using beaker, rather than generated as part of the command execution.
Then the command that's actually run just echos that file.

2. The test utils have been updated such that the timeout for pcp operations
is 6 minutes instead of 1. We don't expect the command to take all 6 minutes,
but the test was failing because the output takes at least 1 minute to
process.