(PE-30624) Deprecate master_uris, add primary_uris
This commit updates pxp-agent as part of the work to deprecate the
`master_uris` parameter. As of this commit, using `master_uris` should
still work, but using the new `primary_uris` should also now work.
pxp-agent will now look for both of those config values when setting
that param.

This commit also updates our tests to use `primary-uris` rather than
`master-uris`, and adds a test to ensure that `master-uris` still works
if `primary-uris` is not present in the config file.

(WIP) Update config