(PE-31871) Extract apply shim to a file
Previously, this shim existed only as a hardcoded Ruby string in a C++
source file. That file would be written when needed and purged after a
period of disuse. But changes to the content weren't necessarily picked
up, as it would only be rewritten after being purged for inactivity.
That led to a problem where pxp-agent installations which had been
upgraded would fail to run apply/apply_prep, because they had an old
copy of the shim. The old version expected to receive a master_uris
argument but pxp-agent had been updated to pass primary_uris instead.

We now manage this file as a regular file and include it during the
build/packaging process so that it will be automatically upgraded by the
package manager.