(maint) Remove exec_prefix_ from BoltModule derived classes
This duplicated the exec_prefix_ field from BoltModule and therefore
calling the parent constructor didn't actually set the field. That
wasn't a problem because most of these classes didn't even use the
field, and the one that did just initialized it explicitly. But either
way that field is redundant so it's removed.