(PE-31216) Disallow null bytes in task output
Previously, pxp-agent would accept task output containing null bytes and
would return it to orchestrator. This caused a database error as null
bytes cannot be written to postgres. These null bytes are also not
allowed by the task spec, so this behavior was incorrect.

The root issue is that the isValidUTF8 helper, which uses rapidjson to
validate each character, considers null bytes to be valid. Technically
this isn't wrong, as the null character is a UTF-8 character, but it's
not a valid character in a null-terminated string.

We now supplement the regular validity check with an additional check
that the string doesn't contain any null bytes.

This also fixes a bug where we could return a result containing invalid
UTF-8 in the case where both stdout and stderr had invalid UTF-8 output.
We would validate that stdout had invalid UTF-8 and would generate an
error including the value from stderr without validating it as UTF-8. We
now simply omit the stderr output as it's unlikely to be useful in this
case.