(PA-4550) Use WMI to lookup windows pid
Several tests rely on pxp-agent to start a process, and then later on, check for
the process and kill it. Prior to cygwin 3.0, the `ps -eW` command returned the
 Windows pid in the PID column.

In Cygwin 3, the `ps` command returns the Cygwin pid and you must use `ps -elW` to
retrieve the Windows pid. From the release notes[1]

    Cygwin PIDs have been decoupled from Windows PID. Cygwin PIDs are now
    incrementally dealt in the range from 2 up to 65535, POSIX-like. Native
    Windows processes not started by Cygwin processes are mapped into the range
    beyond 65535.

In order to find/kill a process started outside of cygwin/bash, like when
pxp-agent starts ping, the test must retrieve the Windows PID and pass that to
taskkill.

This commit eliminates places where we were using `ps` to retrieve Windows pids.
Doing so eliminates the possibility of accidentally matching nssm-pxp-agent.exe
and pxp-agent.exe.

It also fixes a race condition where calling `stop_sleep_process` twice could
fail the second time.

It also fixes a bug when searching for the Ruby.exe process, as the "\s" was not
preserved during ssh/bash evaluation.

[1] https://www.cygwin.com/cygwin-ug-net/ov-new.html#ov-new3.0