(maint) Acceptance Gemfile update
This commit makes the following changes to the acceptance Gemfile:

- Adds rubygems.org as an alternate source.
- Accounts for the deprecated/removed `Object#=~` method in the
location_for method by adding a check to see if the object is a
String.
- Updates the scooter gem, as the version that was pinned (3.2.17)
is dependent on an older version of the json gem (1.8.6) that used
keyword parameter syntax that is deprecated in Ruby 2.7 and is
removed entirely in later Rubies.
- Replaces the `File#exists?` method that has long been deprecated
and was removed entirely in Ruby 3.2 with `File#exist?`.
(See: https://bugs.ruby-lang.org/issues/17391)