Don't capture test.out in pxp-agent.log so it can be archived later
Previously, the test executed `cat /tmp/test.out` via pxp, where test.out
contains 70MB of 'a' characters. The pxp-service uses leatherman to execute
the test and writes the output of the command to pxp-agent.log:

    DEBUG leatherman.execution:93 - executing command: /bin/bash -c cat /tmp/test.out
    DEBUG | - aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa...

The log also contains entries like:

    DEBUG | - {"description":"PXP Puppet module" ... "error":{"type":"string"}

So the post-suite step to print errors to the Jenkins console

    grep -i -B5 -A10 'warn\\|error\\|fatal'" <logfile>

happened to match the above "error" and printed the next 10 lines after
including 70MB of 'a'.

This changes the pxp-agent log level to info so that the contents of test.out
are not captured in the log.