(PE-37481) Return result of downloadFileFromMaster directly on retry
In the original implementation of the retry logic (https://github.com/puppetlabs/pxp-agent/pull/852/files) a while loop using pointers was used to control max retries. On some platforms this approach resulted in broken pointers. Instead, use a for loop and return result of downloadFileFromMaster directly when it does not raise.